// src/app/models/tabDadosGerais.ts
import { StoreType } from "./store";

export class TabDadosGeraisType {
  id: string;
  cnpj: string;
  nome: string;
  situacao: string;
  tipoEstabelecimento: string;
  responsavel: string;
  percentualTaxaIsalon: string;
  flagDestaque: string;
  descricao: string;
  cep: string;
  ufId: string;
  municipioId: string;
  bairroId: string;
  rua: string;
  numeroEndereco: string;
  complementoEndereco: string;
  telefone: string;
  enderecoEmail: string;
  enderecoSite: string;
  diaInicioAtendimentoPadrao: string;
  diaFimAtendimentoPadrao: string;
  horaInicioAtendimentoPadrao: string;
  horaFimAtendimentoPadrao: string;
  diaInicioAtendimentoEspecial: string;
  diaFimAtendimentoEspecial: string;
  horaInicioAtendimentoEspecial: string;
  horaFimAtendimentoEspecial: string;
}